backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
boto3>=1.34
Pillow>=10.0
celery>=5.3
redis>=5.0
requests>=2.31
